### 2020-09-17 #398 @lbusett
presubmission inquiry - prismaread
Submitting Author: Lorenzo Busetto (@lbusett )
Repository:
https://github.com/lbusett/prismaread/
---

```
Package: prismaread
Title: Import PRISMA L1/L2 hyperspectral data and convert them to
 a more user friendly format
Version: 1.0.0
    c(person("Lorenzo", "Busetto", email = "lbusett@gmail.com", role = c("aut", "cre"),
    comment = c(ORCID = '0000-0001-9634-6038')),
    person("Luigi", "Ranghetti", email = "ranghetti.l@irea.cnr.it", role = c("aut"),
    comment = c(ORCID = '0000-0001-6207-5188')))
Description: `prismaread` allows easily importing PRISMA (http://www.prisma-i.it/index.php/it/) satellite
 hyperspectral data cubes and ancillary information from the original data provided by ASI in HDF format,
and convert them to a easier to use format (ENVI or GeoTiff). It also provides functionality for automatically
computing Spectral Indexes from either the original HDF data or from hyperspectral data already converted
using function `pr_convert`, and for easily and quickly extracting  data and computing statistics for the
different bands over areas of interest.
    hdf5r,
    tools,
    raster,
    exactextractr,
    sf,
    tidyselect,
    openxlsx,
    DT,
    assertthat
Suggests: testthat,
    piggyback,
    usethis,
    mapview,
    piggyback
    isciences/exactextractr
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

**data munging** because `prismaread` allows easily importing PRISMA data from the rather complex original HDF format and save them to easier to use raster formats. It also allows to extract ancillary datasets
stored alongside the hyperspectral information (e.g., cloud and land cover masks; acquisition angles, latlon matrixes) and save them as rasters aligned with the hyperspectral data. Finally, it provides functionality for automatically computing Spectral Indexes frome the spectral cubes, and for accessing and storing spectral information over areas of interest.

**geospatial** because `prismaread` focuses on accessing and manipulating geospatial hyperspectral data, ans converting them to easier to use geospatial data formats. The data made available by `prismaread` functionality can be used for geospatial analysis of different kinds.

The target audience is mainly made of Remote Sensing data practitioners interested in testing the capabilities of the innovative PRISMA hyperspectral sensor for analysing the characteristics of the earth surface. Scientific applications of PRISMA data are those typical of satellite hyperspectral imagery, including for example  land cover/use characterization, extraction of biophysical parameters from spectral data, vegetation status monitoring, characterization of coastal waters, etc.

No other packages for performing `prismaread` functionality are currently available to our knowledge. PRISMA data importers are available now in commercial software (e.g., ENVI, ERMapper), but no other OSS solution is currently available to our knowledge.

Yes

1) The testing suite of the package requires accessing rather large original PRISMA datasets. This is currently implemented exploiting package `piggyback` functionality to upload and download/cache test files from the `prismaread` GITHUB repository. This works well but requires a rather long time for downloading data when tests are executed for the first time (test coverage is alreaydy above 90%);

2) No API for directly downloading PRISMA data is currently availble. Users has to ask access to data and register themselves for data access starting at http://prisma-i.it/index.php/en/, and then download data of interest as zip files before being able to exploit `prismaread` functionality.
