### 2020-09-17 #398 @geanders
@lbusett : Thanks so much for your presubmission! We will be discussing if it would be in-scope. As a first question to help our discussion, it sounds like some of the tasks listed for "data munging" are perhaps closer to "data extraction", in terms of parsing scientific datasets. There is some more guidance in the [rOpenSci book](https://devguide.ropensci.org/policies.html#aims-and-scope) on these categories of scope, and in particular this description from "data extraction" could be helpful:

> "Packages that aid in retrieving data from unstructured sources such as text, images and PDFs, as well as parsing scientific data types and outputs from scientific equipment."

Could you clarify a bit any separation between data extraction and data munging tasks in the package?

Also, in terms of comparing with other available software, could you talk about whether there are other more general packages for importing/extracting data from a HDF format, and how your package provides different or improved functionality in terms of working with the PRISMA data compared to using a more general R package alternative for HDF data?
