### 2020-09-18 #398 @lbusett
@geanders

   thanks for your prompt reply! I'll answer inline.

> @lbusett : Thanks so much for your presubmission! We will be discussing if it would be in-scope. As a first question to help our discussion, it sounds like some of the tasks listed for "data munging" are perhaps closer to "data extraction", in terms of parsing scientific datasets. There is some more guidance in the [rOpenSci book](https://devguide.ropensci.org/policies.html#aims-and-scope) on these categories of scope, and in particular this description from "data extraction" could be helpful:
>
> > "Packages that aid in retrieving data from unstructured sources such as text, images and PDFs, as well as parsing scientific data types and outputs from scientific equipment."
>
> Could you clarify a bit any separation between data extraction and data munging tasks in the package?

Yes, after reading the description more carefully I'd guess that the Data Extraction category could fit better, since `prismaread` main objective is to parse/retrieve the data coming from a scientific equipment (the PRISMA hyperspectral sensor) to make it more easily accessible to users. I was probably a bit mislead by the "tools for handling data in specific scientific formats" description of the Data Munging category.

Concerning the separation between "data munging" and "data extraction": The main `prismaread` objective is to access the original PRISMA data, retrieve the different available information and make it available to users as standard and more easily accessible raster datasets (See also the next reply). So, this probably would fit more in the "Data Extraction" categorization. What I termed as "data extraction" functionality in my previous post refers to a function providing a dedicated wrapper to the `exactextract` function of package `exactextracter` (See https://lbusett.github.io/prismaread/articles/Extracting-data-over-vector.html) that can be used to extract original data / statistics over area of interest starting from an already preprocessed PRISMA spectral cube, providing functionality to retrieve it in either long/wide formats and save the retrieved data as RData, excel or CSV files.

> Also, in terms of comparing with other available software, could you talk about whether there are other more general packages for importing/extracting data from a HDF format, and how your package provides different or improved functionality in terms of working with the PRISMA data compared to using a more general R package alternative for HDF data?

Several "R" packages allow general access to HDF data, the "main" ones being `hdf5r`, `raster`, `stars`. However, HDF5 files can be considered as general use containers that can store a variety of types of data and metadata without any strict "rules" about how data is organized within them (see https://support.hdfgroup.org/HDF5/whatishdf5.html - HDF5 is very versatile, but sometimes at the expense of ease of use/access).
Indeed, `prismaread` exploits functionality of some of these more general use packages (mainly, `hdf5r` to access the HDF datasets, and `raster` to reshape / export them) to access the rather complex specific structure of PRISMA raw hdf files. PRISMA data contains infact several subdatasets (e.g., spectral data, ancillary data such as cloud masks or info about acquisition angles ,...), and those subdatasets vary according to the Processing Level of PRISMA data ( L2D = geocoded reflectances, L2C non geocoded reflectances, L2B surface radiance, L1 at sensor radiance ). It also contains metadata attributes that need to be accessed and used, for example, to geocode the images or apply scaling factors to the data. This makes properly importing the available geospatial data a non-trivial task.

In this context `prismaread` makes it painless to the user to extract the datasets of interest and convert them to user friendly formats (See https://lbusett.github.io/prismaread/articles/Importing-Level-1-Data.html; https://lbusett.github.io/prismaread/articles/Importing-Level-2-Data.html).
This requires performing tasks such as matrix transpositions to transform the data cubes to a proper "x,y,wavelength" order, bow/tie geocoding to associate coordinates to L1/2B/2C data, parsing metadata attributes to retrieve wavelength information and scaling factors required to convert raw data to physical units, and "parallel" processing of L1 and L2 data to be able to associate geocoding and acquisition angles information also to L1 data.

With respect to the "general" packages providing access to HDF data, `prismaread` is therefore specifically dedicated to parsing the format of PRISMA data and performing all required operations needed to extract information of interest.

I hope I have sufficiently replied to your questions, but can elaborate more if needed!

   regards,

Lorenzo
