### 2020-09-30 6ae337a
attempt changing NA FLAG
