### 2020-09-30 ae57537
write data ignore value to envi header
