+++ R/pr_create_swir.R
@@ -204,6 +204,7 @@ pr_create_swir <- function(f,
               out_hdr, append = TRUE)
         write("wavelength units = Nanometers")
         write("sensor type = PRISMA")
+        write("data ignore value = -9.99000000e+002")
     }
 
     rm(rast_swir)
