### 2020-09-30 17268d3
write data ignore value and wvl units
