### 2020-09-30 b582cf7
read wvl only from wvl file (tentative)
