### 2020-10-06 #398 @melvidoni
Hello @lbusett. This package is straightforwardly in-scope under data munging and data extraction. We welcome a full submission when it is ready.
