### 2020-12-01 #398 @stefaniebutland
Very sad news. I just learned that Lorenzo Busetto passed away. Beautiful tribute and dedication in the MODIStsp package he co-developed with Luigi Ranghetti: https://docs.ropensci.org/MODIStsp/articles/lorenzo.html
